code13/09

+++ lib/StaffModule/screen/Leasing/RentalRoll/make_payment.dart
@@ -224,7 +224,7 @@ class _MakePaymentState extends State<MakePayment> {
 
   List<Map<String, dynamic>> rows = [];
 
-  List<int> charges_balances = [0];
+  List<double> charges_balances = [0];
 
   //Security Deposite
   // void addRow() {
@@ -431,9 +431,9 @@ class _MakePaymentState extends State<MakePayment> {
             [];
         for (var i = 0; i < filteredCharges!.length; i++) {
           if (i == 0) {
-            charges_balances[0] = filteredCharges[i].chargeAmount!;
+            charges_balances[0] = filteredCharges[i].chargeAmount!.toDouble();
           } else {
-            charges_balances.add(filteredCharges[i].chargeAmount!);
+            charges_balances.add(filteredCharges[i].chargeAmount!.toDouble());
           }
         }
 
@@ -508,7 +508,7 @@ class _MakePaymentState extends State<MakePayment> {
           print(charge);
           print(amount);
           rows[index]['amount'] = amount;
-          charges_balances[index] = (charge.toDouble() + amount).toInt();
+          charges_balances[index] = (charge.toDouble() + amount).toDouble();
           totalAmount += amount;
 
           totalAmount = 0.0;
@@ -524,7 +524,7 @@ class _MakePaymentState extends State<MakePayment> {
           print(charge);
           print(amount);
           rows[index]['amount'] = amount;
-          charges_balances[index] = (charge.toDouble() - amount).toInt();
+          charges_balances[index] = (charge.toDouble() - amount).toDouble();
           totalAmount += amount;
 
           totalAmount = 0.0;
